Pre: Move db/errors.rs into db_traits
