+++ db-traits/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "db_traits"
+version = "0.0.1"
+workspace = ".."
+
+[lib]
+name = "db_traits"
+path = "lib.rs"
+
+[dependencies]
+failure = "0.1.1"
+failure_derive = "0.1.1"
+
+[dependencies.edn]
+path = "../edn"
+
+[dependencies.core_traits]
+path = "../core-traits"
+
+[dependencies.rusqlite]
+version = "0.13"
+features = ["limits"]
