+++ db/src/internal_types.rs
@@ -41,8 +41,8 @@ use edn::entities::{
     TxFunction,
 };
 
-use errors;
-use errors::{
+use db_traits::errors as errors;
+use db_traits::errors::{
     DbErrorKind,
     Result,
 };
