+++ db/src/tx_observer.rs
@@ -39,7 +39,7 @@ use edn::entities::{
     OpType,
 };
 
-use errors::{
+use db_traits::errors::{
     Result,
 };
