+++ query-pull/Cargo.toml
@@ -20,6 +20,9 @@ path = "../core-traits"
 [dependencies.mentat_db]
 path = "../db"
 
+[dependencies.db_traits]
+path = "../db-traits"
+
 [dependencies.mentat_query]
 path = "../query"
